+++ linode/objquota/datasource_test.go
@@ -95,5 +95,4 @@ func TestAccDataSourceObjQuota_basic(t *testing.T) {
 			},
 		},
 	})
-
 }
