+++ linode/vpcsubnet/framework_models.go
@@ -24,8 +24,9 @@ type VPCSubnetModel struct {
 
 func FlattenSubnetLinodeInterface(iface linodego.VPCSubnetLinodeInterface) (types.Object, diag.Diagnostics) {
 	return types.ObjectValue(LinodeInterfaceObjectType.AttrTypes, map[string]attr.Value{
-		"id":     types.Int64Value(int64(iface.ID)),
-		"active": types.BoolValue(iface.Active),
+		"id":        types.Int64Value(int64(iface.ID)),
+		"config_id": types.Int64PointerValue(helper.IntPtrToInt64Ptr(iface.ConfigID)),
+		"active":    types.BoolValue(iface.Active),
 	})
 }
