+++ linode/vpcsubnet/framework_schema_resource.go
@@ -13,8 +13,9 @@ import (
 
 var LinodeInterfaceObjectType = types.ObjectType{
 	AttrTypes: map[string]attr.Type{
-		"id":     types.Int64Type,
-		"active": types.BoolType,
+		"id":        types.Int64Type,
+		"config_id": types.Int64Type,
+		"active":    types.BoolType,
 	},
 }
 
@@ -67,7 +68,6 @@ var frameworkResourceSchema = schema.Schema{
 			Computed:    true,
 			CustomType:  timetypes.RFC3339Type{},
 		},
-
 		"linodes": schema.ListAttribute{
 			Computed:    true,
 			ElementType: LinodeObjectType,
